+++ src/domain/codes/mod.rs
@@ -4,13 +4,13 @@ mod gs1_databar_2d;
 mod pdf417;
 mod qrcode;
 
-#[cfg(feature = "gs1_databar_2d")]
+#[cfg(feature = "codes_2d")]
 pub use gs1_databar_2d::*;
 
-#[cfg(feature = "pdf417")]
+#[cfg(feature = "codes_2d")]
 pub use pdf417::*;
 
-#[cfg(feature = "qrcode")]
+#[cfg(feature = "codes_2d")]
 pub use qrcode::*;
 
 #[cfg(feature = "barcodes")]