+++ src/domain/constants.rs
@@ -76,39 +76,39 @@ pub const GS_BARCODE_WIDTH: &[u8] = &[GS, b'w'];
 pub const GS_BARCODE_PRINT: &[u8] = &[GS, b'k'];
 
 // QR codes
-#[cfg(feature = "qrcode")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D: &[u8] = &[GS, b'(', b'k'];
-#[cfg(feature = "qrcode")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_QRCODE_MODEL: &[u8] = &[GS, b'(', b'k', 4, 0, 49, 65];
-#[cfg(feature = "qrcode")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_QRCODE_SIZE: &[u8] = &[GS, b'(', b'k', 3, 0, 49, 67];
-#[cfg(feature = "qrcode")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_QRCODE_CORRECTION_LEVEL: &[u8] = &[GS, b'(', b'k', 3, 0, 49, 69];
-#[cfg(feature = "qrcode")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_QRCODE_PRINT_SYMBOL_DATA: &[u8] = &[GS, b'(', b'k', 3, 0, 49, 81, 48];
 
 // GS1 DataBar
-#[cfg(feature = "gs1_databar_2d")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_GS1_DATABAR_WIDTH: &[u8] = &[GS, b'(', b'k', 3, 0, 51, 67];
-#[cfg(feature = "gs1_databar_2d")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_GS1_DATABAR_WIDTH_EXTENDED: &[u8] = &[GS, b'(', b'k', 3, 0, 51, 71];
-#[cfg(feature = "gs1_databar_2d")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_GS1_DATABAR_PRINT: &[u8] = &[GS, b'(', b'k', 3, 0, 51, 81, 48];
 
 // PDF417
-#[cfg(feature = "pdf417")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_PDF417_COLUMNS: &[u8] = &[GS, b'(', b'k', 3, 0, 48, 65];
-#[cfg(feature = "pdf417")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_PDF417_ROWS: &[u8] = &[GS, b'(', b'k', 3, 0, 48, 66];
-#[cfg(feature = "pdf417")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_PDF417_WIDTH: &[u8] = &[GS, b'(', b'k', 3, 0, 48, 67];
-#[cfg(feature = "pdf417")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_PDF417_ROW_HEIGHT: &[u8] = &[GS, b'(', b'k', 3, 0, 48, 68];
-#[cfg(feature = "pdf417")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_PDF417_CORRECTION_LEVEL: &[u8] = &[GS, b'(', b'k', 3, 0, 48, 69];
-#[cfg(feature = "pdf417")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_PDF417_TYPE: &[u8] = &[GS, b'(', b'k', 3, 0, 48, 70];
-#[cfg(feature = "pdf417")]
+#[cfg(feature = "codes_2d")]
 pub const GS_2D_PDF417_PRINT: &[u8] = &[GS, b'(', b'k', 3, 0, 48, 81, 48];
 
 // Image