+++ examples/codes.rs
@@ -5,8 +5,8 @@ use escpos::{driver::*, errors::Result};
 fn main() -> Result<()> {
     env_logger::init();
 
-    let driver = NetworkDriver::open("192.168.1.248", 9100)?;
-    // let driver = ConsoleDriver::open(true);
+    // let driver = NetworkDriver::open("192.168.1.248", 9100)?;
+    let driver = ConsoleDriver::open(true);
     let mut printer = Printer::new(driver, Protocol::default());
     printer
         .debug_mode(Some(DebugMode::Dec))
@@ -51,6 +51,9 @@ fn main() -> Result<()> {
         // MaxiCode
         .writeln("MaxiCode")?
         .maxi_code_option("1245789658745", MaxiCodeMode::Mode2)?
+        // DataMatrix
+        .writeln("DataMatrix")?
+        .data_matrix("test1245789658745")?
         .feed()?
         .print_cut()?;